git-svn-id: https://svn.code.sf.net/p/teamengine/code/branches/team2@305 24c30046-25e5-4eec-8353-25991c8b1332

+++ apps/engine/src/java/com/occamlab/te/util/Misc.java
@@ -44,16 +44,22 @@ public class Misc {
 
     // Deletes a directory and its contents
     public static void deleteDir(File dir) {
+        if (dir.isDirectory()) {
+            deleteDirContents(dir);
+            dir.delete();
+        }
+    }
+
+    // Deletes the contents of a directory
+    public static void deleteDirContents(File dir) {
         String[] children = dir.list();
         for (int i = 0; i < children.length; i++) {
             File f = new File(dir, children[i]);
             if (f.isDirectory()) {
-                deleteDir(f);
-            } else {
-                f.delete();
+                deleteDirContents(f);
             }
+            f.delete();
         }
-        dir.delete();
     }
 
     // Deletes just the sub directories for a certain directory
